pyproject.toml
[project]
name = "test-nats"
version = "0.1.0"
description = "NATS concept testing environment"
dependencies = [
    "nats-py>=2.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
